requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
requests
python-whois
beautifulsoup4
